M:1 with nested tasks
This snapshot represents a worked example of a `yield()` operator that
respects the "stack" of tasks, which may more naturally map to
expectations for how control ought to flow.

I'm also intending to pause on this toy example for now, since pthreads
are a whole _thing_ and I'm fairly confident I've learned what I needed
to without going to them.